move preloadAtPublicGateway option to file import section
